requirements.txt
bleak==0.22.3
bluepy==1.3.0
google-cloud-pubsub==2.27.1
google-cloud-bigquery==3.27.0
functions-framework==3.8.2
google-cloud-secret-manager==2.21.1
cloudevents==1.11.0
requests==2.32.3
geojson==3.1.0